# Improve perf via recursion