--- a/ProtoPromise_Unity/Assets/Plugins/ProtoPromise/Core/Promises/Internal/PromiseFieldsInternal.cs
+++ b/ProtoPromise_Unity/Assets/Plugins/ProtoPromise/Core/Promises/Internal/PromiseFieldsInternal.cs
@@ -483,7 +483,6 @@ namespace Proto.Promises
 
             partial class AsyncPromiseRef<TResult> : AsyncPromiseBase<TResult>
             {
-                private HandleablePromiseBase _nextForComplete;
                 private ExecutionContext _executionContext;
 #if PROMISE_PROGRESS
                 private float _minProgress;