Since #99 made it possible to force async continuations, we no longer need to unwind the stack before invoking continuations. If users have very deep async stacks, they can use the `forceAsync` flag on `WaitAsync` or `SwitchToContext` / `SwitchToForeground` / `SwitchToBackground` APIs to prevent stack overflows (or `Task.Yield()` in async functions). This gains us ~7% performance and slightly reduced memory.

Master:

|         Type | Pending |       Mean | Allocated | Survived |
|------------- |-------- |-----------:|----------:|---------:|
|   AsyncAwait |    True | 2,385.6 ns |         - |    720 B |
| ContinueWith |    True | 2,979.8 ns |         - |    400 B |

This PR:

|         Type | Pending |       Mean | Allocated | Survived |
|------------- |-------- |-----------:|----------:|---------:|
|   AsyncAwait |    True | 2,224.6 ns |         - |    696 B |
| ContinueWith |    True | 2,738.8 ns |         - |    400 B |